+++ src/components/CheckboxIOS.js
@@ -34,7 +34,7 @@ type Props = {
 /**
  * Checkboxes allow the selection of multiple options from a set.
  */
-class Checkbox extends React.Component<Props> {
+class CheckboxIOS extends React.Component<Props> {
   render() {
     const { checked, disabled, onPress, theme, ...rest } = this.props;
 
@@ -88,4 +88,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default withTheme(Checkbox);
+export default withTheme(CheckboxIOS);
